Successfully moved everything into one vue file
